refactor: isolate world pipeline and remove stale seams
