+++ .gitignore
@@ -37,3 +37,4 @@ logs/
 build/
 dist/
 out/
+graphify-out/
