+++ manifest.json
@@ -25,5 +25,5 @@
   ],
   "affects_link": false,
   "github": "ShaneMcGovernIE/potato_voxel",
-  "description": "PotatoVoxel (fork of Dramatic Shape): the overworld as a 3D voxel diorama, retuned for low-end devices. Same diorama, with the settings that would otherwise cost a handheld its frame budget: the VOXEL ladder is a set of quality MODES (OFF / HIGH / MEDIUM / LOW / POTATO / CUSTOM) -- picking a mode applies its tuned defaults to the quality knobs (WATER, AA, V-CURVE, V-GRID, 3D-BTL, SHADOWS), and touching any knob flips the mode to CUSTOM. A RENDER SCALE knob (100% / 75% / 50% / 33%) is the single biggest frame-budget lever, set by the modes and adjustable on its own; scoped mesh-cache storage. One build for every device -- no desktop environment switch. 1.6.1 runs under the mod sandbox: no raw file access, no OS queries, and the Stadium/VR/Forest FX features that needed them are gone."
+  "description": "PotatoVoxel (fork of Dramatic Shape): the overworld as a 3D voxel diorama, retuned for low-end devices. Same diorama, with the settings that would otherwise cost a handheld its frame budget: the VOXEL ladder is a set of quality MODES (OFF / HIGH / MEDIUM / LOW / POTATO / CUSTOM) -- picking a mode applies its tuned defaults to the quality knobs (WATER, AA, V-CURVE, V-GRID, 3D-BTL, SHADOWS), and touching any knob flips the mode to CUSTOM. A RENDER SCALE knob (100% / 75% / 50% / 33%) is the single biggest frame-budget lever, set by the modes and adjustable on its own; scoped mesh-cache storage. One build for every device -- no desktop environment switch. 1.7.11 runs under the mod sandbox: no raw file access, no OS queries, and the Stadium/VR/Forest FX features that needed them are gone."
 }
