+++ mod.card
@@ -8,7 +8,7 @@ return {
            "battle", "low-end" },
   differences = {
     changed = {
-      "PotatoVoxel (fork of Dramatic Shape Voxel Mod, formerly the TrimUI Brick build): the low-end tuning is the DEFAULT everywhere, and the VOXEL ladder is a set of quality MODES -- OFF / HIGH / MEDIUM / LOW / POTATO / CUSTOM. Picking a mode applies its tuned defaults to every quality knob (WATER, AA, V-CURVE, V-GRID, 3D-BTL, SHADOWS) and to the RENDER SCALE knob (100% / 75% / 50% / 33%, folded back with ordinary canvas filtering; no dedicated upscaling shader or RCAS pass); touching any knob flips the mode to CUSTOM. Desktop HIGH keeps the full two-layer actor shadow map, including moving actor/sprite casters; MEDIUM and lower use contact/blob actor shadows. Every quality knob is tuned for a quad-A53/PowerVR/998MB device and DEFAULTS OFF -- WATER, AA, V-CURVE, V-GRID, 3D-BTL (flat card battles) -- but each is a switchable row on the VOXEL SETTINGS submenu, and the voxel mesh build budget is halved per frame so chunk carving yields to gameplay on four cores. There is one build for every device -- no desktop environment switch; the mesh cache lives in the mod sandbox's scoped storage. 1.6.1 drops what the sandbox forbids: FOREST FX, the frosted battle-HUD glass, the Stadium model import (player-supplied ROM reads are impossible), VR, and the diagnostics panel are gone -- shadows now run on every device that can do them.",
+      "PotatoVoxel (fork of Dramatic Shape Voxel Mod, formerly the TrimUI Brick build): the low-end tuning is the DEFAULT everywhere, and the VOXEL ladder is a set of quality MODES -- OFF / HIGH / MEDIUM / LOW / POTATO / CUSTOM. Picking a mode applies its tuned defaults to every quality knob (WATER, AA, V-CURVE, V-GRID, 3D-BTL, SHADOWS) and to the RENDER SCALE knob (100% / 75% / 50% / 33%, folded back with ordinary canvas filtering; no dedicated upscaling shader or RCAS pass); touching any knob flips the mode to CUSTOM. Desktop HIGH keeps the full two-layer actor shadow map, including moving actor/sprite casters; MEDIUM and lower use contact/blob actor shadows. Every quality knob is tuned for a quad-A53/PowerVR/998MB device and DEFAULTS OFF -- WATER, AA, V-CURVE, V-GRID, 3D-BTL (flat card battles) -- but each is a switchable row on the VOXEL SETTINGS submenu, and the voxel mesh build budget is halved per frame so chunk carving yields to gameplay on four cores. There is one build for every device -- no desktop environment switch; the mesh cache lives in the mod sandbox's scoped storage. 1.7.11 runs under the mod sandbox: FOREST FX, the frosted battle-HUD glass, Stadium model import, and VR are removed; bounded diagnostics remain sandbox-safe, and shadows run on every device that can do them.",
       "with VOXEL on, the overworld draws as 3D geometry instead of flat tiles",
       "occlusion comes from a depth buffer rather than a y-sort, so buildings really hide what is behind them",
       "with 3D-BTL on 2D-3D A, a battle draws over the map's nearest clear ground instead of over a white field",
@@ -49,5 +49,5 @@ return {
     { who = "pret/pokered", for_ = "the tile and sprite data the geometry is derived from" },
     { who = "AverageConsumer", for_ = "battle UI/sprite interoperability (honouring the engine's visibility seams and composing staged-battle sprite overrides) and the cold mesh-build Town Map loading cover" },
   },
-  compat = { engine = ">=0.1.37 <1.6.1", modApi = 2 },
+  compat = { engine = ">=0.1.37 <2.0.0", modApi = 2 },
 }
